+++ lib/liquid_diagrams/renderers/blockdiag_renderer.rb
@@ -2,12 +2,7 @@
 
 module LiquidDiagrams
   module Renderers
-    %i[ Blockdiag
-        Seqdiag
-        Actdiag
-        Nwdiag
-        Rackdiag
-        Packetdiag ].each do |diagram|
+    %i[Blockdiag Seqdiag Actdiag Nwdiag Rackdiag Packetdiag].each do |diagram|
       renderer = Class.new(BasicRenderer) do
         const_set :OPTIONS, %w[
           config
@@ -29,14 +24,11 @@ module LiquidDiagrams
         define_method :build_command do
           command = +"#{diagram.downcase} -T svg --nodoctype"
 
-          options = self.class.const_get(:OPTIONS)
-          switches = self.class.const_get(:SWITCHES)
-
-          @config.slice(*options).each do |opt, value|
+          @config.slice(*self.class.const_get(:OPTIONS)).each do |opt, value|
             command << " --#{opt}=#{value}"
           end
 
-          Utils.merge(switches, @config).each do |swc, value|
+          Utils.merge(self.class.const_get(:SWITCHES), @config).each do |swc, value|
             command << " --#{swc}" if value
           end