+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.2.0 - 2020.05.02
+
+### Features
+
+* Add `vegalite`
+
+### Bugfix
+
+* Fix `plantuml` and `smcat` case error
+* Fix template parse options
+
+### Refactor
+
+* Optimize method call chain in `BasicBlock`
+
 ## 0.1.0 - 2020.05.02
 
 ### Features